Update CustomTNTFlow namespace
