+++ stubs/customtntflow/api-src/dev/byflow/customtntflow/api/RegionTNTType.java
@@ -1,4 +1,4 @@
-package dev.byflow.customtnt.api;
+package dev.byflow.customtntflow.api;
 
 public interface RegionTNTType {
     String getId();
